# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company API",
        "version": "1.0"
    },
    "paths": {
        "/companies/building/{buildingId}": {
            "get": {
                "summary": "Вы можете получить компании по здании",
                "description": "Возвращает список компаний, расположенных в определенном здании.",
                "operationId": "adf889682fc6da158fb2380f3994a38e",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список компаний в указанном здании.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/companies/category/{categoryId}": {
            "get": {
                "summary": "Вы можете получить компании по категориям",
                "description": "Возвращает список компаний в определенной категории.",
                "operationId": "c0774722b105f97b50a4a31bb1d0ddec",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список компаний в указанной категории.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/companies/location": {
            "get": {
                "summary": "Вы можете получить компании по местоположению и радиусу",
                "description": "Возвращает список компаний, расположенных в пределах заданного радиуса или прямоугольной области относительно заданной точки на карте.",
                "operationId": "49f5f0e7d21d6d67b39b27342c528079",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список компаний по местоположению.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "summary": "Вы можете получить компанию по ID",
                "description": "Возвращает информацию о компании по ID.",
                "operationId": "5ec6b00efeefe84dca1930d6f1ff99f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о компании.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/companies/search/activity": {
            "get": {
                "summary": "Поиск компаний по виду деятельности",
                "description": "Ищет компании по типу деятельности, возвращая все компании, соответствующие данному виду деятельности, и соответствующие подкатегории.",
                "operationId": "7173ae454dbf977b6ca6d7b8891b72ed",
                "parameters": [
                    {
                        "name": "activity",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список компаний, соответствующих запросу по виду деятельности.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/companies/search/name": {
            "get": {
                "summary": "Поиск компаний по названию",
                "description": "Ищет компании по названию.",
                "operationId": "3128e1dac9d451269e11e00d5521392d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список компаний, соответствующих запросу по названию.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Company"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Building": {
                "required": [
                    "id",
                    "address",
                    "company_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "1234 Street Name"
                    },
                    "coordinates": {
                        "type": "array",
                        "items": {
                            "type": "float"
                        },
                        "example": {
                            "lat": 40.7128,
                            "lng": -74.006
                        }
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    }
                },
                "type": "object"
            },
            "Category": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Tech"
                    },
                    "parent": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "companies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Company"
                        }
                    }
                },
                "type": "object"
            },
            "Company": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Company A"
                    },
                    "buildings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Building"
                        }
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "CompanyActivity": {
                "required": [
                    "company_id",
                    "category_id"
                ],
                "properties": {
                    "company_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "Phone": {
                "required": [
                    "id",
                    "company_id",
                    "phone_number"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    }
                },
                "type": "object"
            }
        }
    }
}